+++ FMCore/LFO.cs
@@ -7,42 +7,42 @@ public class LFO : Resource //: Node
 {
     public const string _iotype = "lfo";
     // This value should typically be initialized to whatever the global sample rate is.  Require it in the ctor
-    public static double sample_rate = 44100.0;
+    public static float sample_rate = 44100.0f;
     public bool enabled = false;  //Used by Patch to determine whether it should update the timer and recalculate the phase position from an oscillator.
 
     public int id=0;  //LFO bank number.  FIXME:  Needed?
 
-    private double[] buffer=new double[1];
+    private float[] buffer=new float[1];
 
 
     //User-specified values
     public Waveforms waveform = Waveforms.SINE;
     public bool reflect_waveform;
 
-    private double depth = 1.0; //The maximum amount this LFO oscillates the frequency multiplier applied to the Operator output. Always positive.
-    public double cycleTime = 1;  //Length, in seconds to reach one oscillation of the LFO waveform.
+    private float depth = 1.0f; //The maximum amount this LFO oscillates the frequency multiplier applied to the Operator output. Always positive.
+    public float cycleTime = 1;  //Length, in seconds to reach one oscillation of the LFO waveform.
     public int delay = 0;  //Delay, in samples, before the oscillation kicks in.
 
-    public double bias = 0;  //DC offset bias
+    public float bias = 0;  //DC offset bias
 
     //Helper properties that allow specifying cycle time and delay in millisecs.
-    public double CycleTime { get =>  cycleTime * 1000; set => cycleTime = (value/1000.0); }
-    public double Delay { get =>  1000 * delay / sample_rate ; set =>  delay = (int)((value/1000.0)*sample_rate); }
-    public double Depth { get => depth*100; set => depth = value/100.0; }
-    public double Bias { get => bias*100; set => bias = value/100.0; }
+    public float CycleTime { get =>  cycleTime * 1000; set => cycleTime = (value/1000.0f); }
+    public float Delay { get =>  1000 * delay / sample_rate ; set =>  delay = (int)((value/1000.0f)*sample_rate); }
+    public float Depth { get => depth*100; set => depth = value/100.0f; }
+    public float Bias { get => bias*100; set => bias = value/100.0f; }
 
 
     //Running counters
     private int samples = 0;  //Samples elapsed.  Sample timer.  Can be reset by key sync.
-    public double phase = 0.0;  //Phase accumulator for this LFO.
-    public double currentOscillation = 1.0;  //This value is precalculated whenever the sample timer changes.  Reference it when applying totalMultiplier to save CPU.
+    public float phase = 0.0f;  //Phase accumulator for this LFO.
+    public float currentOscillation = 1.0f;  //This value is precalculated whenever the sample timer changes.  Reference it when applying totalMultiplier to save CPU.
 
     //Extra Options
     public bool keySync=true;  //Affects resets on the sample timer when a new key is pressed; if disabled, delay is only used to affect an LFO's phase offset.
     public bool oscSync=true;  //Affects resets on the accumulator when a new key is pressed, otherwise the oscillator continues to cycle it from previous position.
     public bool legato = true;  //If disabled, quantizes LFO values to nearest 12th root of 2
 
-    // public static readonly double[] NEAREST_12 = new double[] //Nearest power of 12, used for quantizing LFO when discrete note (no legato) mode is enabled
+    // public static readonly float[] NEAREST_12 = new float[] //Nearest power of 12, used for quantizing LFO when discrete note (no legato) mode is enabled
     //         {0, 0.059463, 0.122462, 0.189207, 0.259921, 0.33484, 0.414214, 0.498307, 0.587401, 0.681793, 0.781797, 0.887749, 1.0};
 
     #region IO
@@ -113,7 +113,7 @@ public class LFO : Resource //: Node
 
     //Constructors
     public LFO(){}
-    public LFO(double sample_rate)
+    public LFO(float sample_rate)
     {
         LFO.sample_rate = sample_rate;
     }
@@ -121,14 +121,14 @@ public class LFO : Resource //: Node
 
     /// Sets this LFO's buffer to the next chunk of frames needed by the audio buffer filling algorithm. Each frame is used as a reference by operators.
     public void UpdateBuffer(int length) {buffer = RequestBuffer(length);}
-    public double[] RequestBuffer(int length)
+    public float[] RequestBuffer(int length)
     {
-        var output = new double[length];
+        var output = new float[length];
 
         bool delayed;  //Check if we need for delay to expire
         for (int i=0; i < length; i++)
         {
-            if (!enabled) {output[i] = 1.0; continue;}
+            if (!enabled) {output[i] = 1.0f; continue;}
                 else {output[i] = Calc(out delayed);}
 
             Iterate();
@@ -141,17 +141,17 @@ public class LFO : Resource //: Node
     public virtual void Reset(bool force = false)
     {
         if (keySync || force)  samples=0;
-        if (oscSync || force){ phase = 0.0;  
+        if (oscSync || force){ phase = 0.0f;  
             if (!keySync)  Accumulate(delay); } //When there's oscillator sync but no key sync, presume the delay specifies a phase offset.
         
     }
 
     /// Calculates a multiplier at the current LFO's sample position.  Used for Operators' pitch modulation sensitivity.
-    public virtual double GetPitchMult(int noteSamples, double sensitivity=1.0)
+    public virtual float GetPitchMult(int noteSamples, float sensitivity=1.0f)
     {
         if (samples < delay)
         {
-            return 1.0;  //No adjustment to the LFO multiplier.  Wait until delay's passed.
+            return 1.0f;  //No adjustment to the LFO multiplier.  Wait until delay's passed.
         } else {
             //First, apply the sensitivity to the output before converting it to a pitch multiplier.
             //TODO:  Should depth magnitude also be applied here?  
@@ -163,18 +163,18 @@ public class LFO : Resource //: Node
             //     output =  1 / Math.Abs(1-output); 
             // } else { output += 1; }
 
-            output = Math.Pow(2, output);
+            output = (float) Math.Pow(2, output);
 
             return output+bias;  
         }
     }
 
     /// Calculates a multiplier at the specified buffer position.  Used for Operators' pitch modulation sensitivity.
-    public virtual double GetPitchMult(int idx, int noteSamples, double sensitivity=1.0)
+    public virtual float GetPitchMult(int idx, int noteSamples, float sensitivity=1.0f)
     {
         if (samples < delay)  //Determine if the note is mature enough to have LFO applied.
         {
-            return 1.0;  //No adjustment to the LFO multiplier.  Wait until delay's passed.
+            return 1.0f;  //No adjustment to the LFO multiplier.  Wait until delay's passed.
         } else {
             var osc = buffer[idx];
             var output = depth * osc * sensitivity ;
@@ -185,7 +185,7 @@ public class LFO : Resource //: Node
             //     output =  1 / Math.Abs(1-output); 
             // } else { output += 1; }
 
-            output = Math.Pow(2, output);
+            output = (float) Math.Pow(2, output);
 
             return output+bias;  
         }
@@ -193,34 +193,34 @@ public class LFO : Resource //: Node
 
 
     /// Calculates a multiplier at the current LFO's sample position.  Used for Operators' total amplitude modulation sensitivity.
-    public virtual double GetAmpMult(int noteSamples, double sensitivity=1.0)
+    public virtual float GetAmpMult(int noteSamples, float sensitivity=1.0f)
     {
         if (samples < delay)
         {
-            return 1.0;  //No adjustment to the LFO multiplier.  Wait until delay's passed.
+            return 1.0f;  //No adjustment to the LFO multiplier.  Wait until delay's passed.
         } else {
             //The minimum level of amplitude is determined by the sensitivity.  This will be added to the output; the final value will be normalized.
             //The result is that a sensitivity of 0 always keeps output TL at max, and a full sensitivity of 1 will oscillate output from 0-1.            
-            var output = (currentOscillation * sensitivity) + (1.0-sensitivity);
+            var output = (currentOscillation * sensitivity) + (1.0f-sensitivity);
 
             //Output can range from -1 to 1 at this point, so scale it to the proper level for amplitude output.
-            return (output+1) * 0.5;
+            return (output+1) * 0.5f;
         }
     }
     /// Calculates a multiplier at the specified buffer position.  Used for Operators' total amplitude modulation sensitivity.
-    public virtual double GetAmpMult(int idx, List<double> ampBuffer, double sensitivity=1.0)
+    public virtual float GetAmpMult(int idx, List<float> ampBuffer, float sensitivity=1.0f)
     {
-        double output;
+        float output;
         if (samples < delay)
         {
-            output = 1.0;  //No adjustment to the LFO multiplier.  Wait until delay's passed.
+            output = 1.0f;  //No adjustment to the LFO multiplier.  Wait until delay's passed.
         } else {
             //The minimum level of amplitude is determined by the sensitivity.  This will be added to the output; the final value will be normalized.
             //The result is that a sensitivity of 0 always keeps output TL at max, and a full sensitivity of 1 will oscillate output from 0-1.            
-            output = (buffer[idx] * sensitivity) + (1.0-sensitivity);
-            output = (output +1) * 0.5;
+            output = (buffer[idx] * sensitivity) + (1.0f-sensitivity);
+            output = (output +1) * 0.5f;
         }
-            var average = 0.0;
+            var average = 0.0f;
             for (int i=0; i < ampBuffer.Count; i++){
                 average += ampBuffer[i];
             }   average /= ampBuffer.Count;
@@ -235,12 +235,12 @@ public class LFO : Resource //: Node
 
 
     /// Gets the oscillator at current phase.
-    public double Calc(out bool delayed)
+    public float Calc(out bool delayed)
     {
         if (samples < delay)
         {
             delayed = true;
-            return 0.0;
+            return 0.0f;
         } else {
             delayed = false;
             var output = oscillators.wave(phase, waveform, reflect: reflect_waveform);
@@ -256,7 +256,7 @@ public class LFO : Resource //: Node
     }
 
     //Is any multiplier != 1.0 necessary?  FIXME later
-    public void Accumulate(int numsamples=1, double multiplier=1)
+    public void Accumulate(int numsamples=1, float multiplier=1)
     {
 
         phase += numsamples / sample_rate / cycleTime ;
@@ -270,7 +270,7 @@ public class LFO : Resource //: Node
     }
 
     /// Retreives the nearest semitone to a given input position (-1.0 to 1.0)
-    static double NearestSemitone(double input)
+    static float NearestSemitone(double input)
     {
         // var sign = Math.Sign(input);
         // var whole = Math.Truncate(input);
@@ -279,7 +279,7 @@ public class LFO : Resource //: Node
         // var idx = (int) (frac * (NEAREST_12.Length-double.Epsilon));
         // return (NEAREST_12[idx] + (whole>1? whole:0)) * sign ;
 
-        return Math.Round(input*12)/12;
+        return (float) Math.Round(input*12)/12;
 
     }
 }
